chore: 调整 Electron 目录，使用新版的 TRTCSimpleDemo 替换旧版，旧版暂时移入 TRTCScenesDemo ，Electron 目录下的 README.md 文件改为和其它端一致。
