+++ Electron/TRTCScenesDemo/resources/gift_sent.aac
