+++ Electron/TRTCScenesDemo/resources/watermark.png
