+++ Electron/TRTCSimpleDemo/src/app.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <router-view/>
+  </div>
+</template>
