+++ Electron/TRTCSimpleDemo/src/common.css
@@ -1,0 +1,57 @@
+body {
+    background: #fefefe;
+    color: #222;
+    overflow: hidden;
+}
+
+
+#background-img {
+    position: fixed;
+    right: 0;
+    bottom: 0;
+    min-width: 100%;
+    min-height: 100%; 
+    width: auto;
+    height: auto; 
+    z-index: -100;
+    filter: brightness(0.5) blur(15px);
+}
+
+#video-container {
+    width: 100vw;
+    height: 100vh;
+    position: absolute;
+    top: 0;
+    left: 0;
+    z-index: -1;
+  }
+  
+  #video-container > .user-video-container {
+      display: block;
+      position: absolute;
+      overflow: hidden;
+  }
+  
+  #video-container > .user-video-container > div {
+    width: inherit;
+    height:inherit;
+  }
+  
+  #video-container > .local-video-container {
+    width: 100%;
+    height: 100%;
+  }
+  
+  #video-container > .local-video-container>div>canvas{
+    margin:0;
+  }
+
+  .right {
+    position: absolute;
+    right: 10px;
+  }
+  
+  .left {
+    position: absolute;
+    left: 10px;
+  }
