+++ Electron/TRTCSimpleDemo/src/pages/404.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <h1> --404 Not found-- </h1>
+  </div>
+</template>
