+++ Electron/TRTCSimpleDemo/src/pages/index.vue
@@ -1,0 +1,28 @@
+<template>
+  <div>
+    <main-menu> </main-menu>
+  </div>
+</template>
+
+<script>
+
+export default {
+
+}
+</script>
+
+<style scoped>
+#background-img {
+    position: fixed;
+    top: -10%;
+    left: -10%;
+    right: 0px;
+    min-width: 120%;
+    min-height: 130%; 
+    width: auto;
+    height: auto; 
+    z-index: -100;
+    filter: brightness(0.5);
+    margin:auto;
+}
+</style>
